docs: split execution from call
